Restructure gfx_device module
