+++ src/gfx_device/assets/mod.rs
@@ -1,0 +1,5 @@
+pub mod texture;
+pub mod mesh;
+
+pub use self::texture::Texture;
+pub use self::mesh::Mesh;
