+++ src/gfx_device/display_config.rs
@@ -1,0 +1,18 @@
+use config::Element;
+use std::path::Path;
+
+config!(
+    /// Contains display config,
+    /// it is required to call video_init()
+    struct DisplayConfig {
+        pub title: String = "Amethyst game".to_string(),
+        pub fullscreen: bool = false,
+        pub dimensions: Option<(u32, u32)> = None,
+        pub min_dimensions: Option<(u32, u32)> = None,
+        pub max_dimensions: Option<(u32, u32)> = None,
+        pub vsync: bool = true,
+        pub multisampling: u16 = 1,
+        pub visibility: bool = true,
+        pub backend: String = "Null".to_string(),
+    }
+);
