+++ src/gfx_device/world_resources/mod.rs
@@ -1,0 +1,5 @@
+pub mod camera;
+pub mod screen_dimensions;
+
+pub use self::camera::Camera;
+pub use self::screen_dimensions::ScreenDimensions;
